// client/src/app/shared/operation.model.ts
export class Operation {
    cid: number;
    add_id: number;
    order_id:number;
    cname: string;
    ph1:string;
    ph2:string;
    houseNo: string;
    floor_room:string;
    verify: boolean;

    zone: string;
    township: string;
    ward: string;
    street: string;

    // order_id: number;   //number
    orderPoster: string;
    planType: string;
    cpeType: string;
    useType: string;
    orderChannel: string;
    how_u_know: string;
    remark:string;
    status:string;
    toInstall:boolean;
}